Les valeurs de date en Java ont pour type "long", au lieu de "int", pour les Obsel
git-svn-id: https://svn.liris.cnrs.fr/dcram/dev/ktbs@728 2a3803fa-5b2e-0410-8379-b254d7af8764

+++ ktbs-java-core/src/org/liris/ktbs/core/Obsel.java
@@ -11,13 +11,13 @@ public interface Obsel extends KtbsResource {
 	 * 
 	 * @return the relative begin date in milliseconds, -1 if unset
 	 */
-	public int getBegin();
+	public long getBegin();
 
 	/**
 	 * 
 	 * @return the relative end date in milliseconds, -1 if unset
 	 */
-	public int getEnd();
+	public long getEnd();
 
 	public Date getBeginDT();
 	public Date getEndDT();
